+++ test/feature/conditional-event-feature.js
@@ -1,6 +1,11 @@
 import { Definition } from '../../src/definition/Definition.js';
 import testHelpers from '../helpers/testHelpers.js';
 import factory from '../helpers/factory.js';
+import js from '../resources/extensions/JsExtension.js';
+
+const extensions = {
+  js,
+};
 
 const boundJsEventSource = factory.resource('conditional-bound-js-event.bpmn');
 
@@ -58,15 +63,19 @@ Feature('Conditional event', () => {
       expect(definition.getActivityById('conditionalEvent').counters).to.deep.equal({ taken: 0, discarded: 1 });
     });
 
-    let waiting;
+    let wait;
     When('ran again', () => {
       completed = definition.waitFor('leave');
-      waiting = definition.waitFor('wait');
+      wait = definition.waitFor('wait');
       definition.run();
     });
 
-    And('condition is signalled', async () => {
-      await waiting;
+    Then('bound condition is waiting', async () => {
+      const event = await wait;
+      expect(event.content).to.have.property('condition', 'expression');
+    });
+
+    When('condition is signalled', () => {
       definition.signal({ id: 'conditionalEvent' });
     });
 
@@ -213,6 +222,183 @@ Feature('Conditional event', () => {
     });
   });
 
+  Scenario('recover resume', () => {
+    let context, definition;
+    Given('a process with bound javascript condition', async () => {
+      const source = `
+        <definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+          <process id="theProcess" isExecutable="true">
+            <startEvent id="start" />
+            <manualTask id="task" />
+            <boundaryEvent id="cond" attachedToRef="task" cancelActivity="true">
+              <conditionalEventDefinition>
+                <condition xsi:type="tFormalExpression" language="js">
+                  next(null, properties.type === 'signal' && content.message.type === "create");
+                </condition>
+              </conditionalEventDefinition>
+            </boundaryEvent>
+            <endEvent id="end" />
+            <sequenceFlow id="flow0" sourceRef="start" targetRef="task" />
+            <sequenceFlow id="flow1" sourceRef="task" targetRef="end" />
+            <sequenceFlow id="flow2" sourceRef="cond" targetRef="end" />
+          </process>
+        </definitions>`;
+
+      context = await testHelpers.context(source);
+
+      definition = new Definition(context);
+    });
+
+    let state, stopped;
+    When('run and saving state on activity condition', () => {
+      definition = new Definition(context.clone());
+      definition.waitFor('activity.condition', () => {
+        state = definition.getState();
+        definition.stop();
+      });
+      stopped = definition.waitFor('stop');
+
+      definition.run();
+    });
+
+    Then('state is saved', async () => {
+      await stopped;
+    });
+
+    When('recovered and resumed', () => {
+      definition = new Definition(context.clone()).recover(state);
+      definition.resume();
+
+      definition.waitFor('activity.condition', () => {
+        state = definition.getState();
+        definition.stop();
+      });
+      stopped = definition.waitFor('stop');
+    });
+
+    And('condition is signalled with non-matching message', () => {
+      definition.signal({ id: 'cond' });
+    });
+
+    Then('run is stopped', () => {
+      return stopped;
+    });
+
+    let completed;
+    When('recovered and resumed from signal', () => {
+      definition = new Definition(context.clone()).recover(state);
+      definition.resume();
+
+      completed = definition.waitFor('leave');
+    });
+
+    And('condition is signalled with matching message', () => {
+      definition.signal({ id: 'cond', type: 'create' });
+    });
+
+    Then('run completes', () => {
+      return completed;
+    });
+
+    When('recovered and resumed from signal state again', () => {
+      definition = new Definition(context.clone()).recover(state);
+      definition.resume();
+
+      stopped = definition.waitFor('stop');
+    });
+
+    And('condition is signalled with non-matching message and stopped on activity.condition', () => {
+      definition.getActivityById('cond').broker.subscribeOnce('event', 'activity.condition', () => {
+        definition.stop();
+      });
+
+      definition.signal({ id: 'cond' });
+    });
+
+    Then('run is stopped', () => {
+      return stopped;
+    });
+
+    When('run is resumed', () => {
+      definition.resume();
+      completed = definition.waitFor('leave');
+    });
+
+    And('condition is signalled with matching message', () => {
+      definition.signal({ id: 'cond', type: 'create' });
+    });
+
+    Then('resumed run completes', () => {
+      return completed;
+    });
+
+    When('recovered and resumed from signal state yet again', () => {
+      definition = new Definition(context.clone()).recover(state);
+      definition.resume();
+
+      completed = definition.waitFor('leave');
+    });
+
+    And('attached task completes', () => {
+      definition.signal({ id: 'task' });
+    });
+
+    Then('resumed run completes', () => {
+      return completed;
+    });
+  });
+
+  Scenario('signalling', () => {
+    let context, definition;
+    Given('a process matching scenario', async () => {
+      const source = `
+        <definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+          <process id="theProcess" isExecutable="true">
+            <startEvent id="start" />
+            <manualTask id="task" />
+            <boundaryEvent id="cond" attachedToRef="task" cancelActivity="true">
+              <conditionalEventDefinition>
+                <condition xsi:type="tFormalExpression" language="js">
+                  next(null, properties.type === 'signal' && content.message.type === "create");
+                </condition>
+              </conditionalEventDefinition>
+            </boundaryEvent>
+            <endEvent id="end" />
+            <sequenceFlow id="flow0" sourceRef="start" targetRef="task" />
+            <sequenceFlow id="flow1" sourceRef="task" targetRef="end" />
+            <sequenceFlow id="flow2" sourceRef="cond" targetRef="end" />
+          </process>
+        </definitions>`;
+
+      context = await testHelpers.context(source);
+
+      definition = new Definition(context);
+    });
+
+    let wait, completed;
+    When('ran waiting for signal', () => {
+      definition = new Definition(context.clone());
+      wait = definition.waitFor('wait', (_, msg) => {
+        return msg.content.id === 'cond';
+      });
+      completed = definition.waitFor('leave');
+
+      definition.run();
+    });
+
+    And('signalled multiple times where one in the middle matches condition', async () => {
+      await wait;
+
+      for (let i = 0; i < 15; i++) {
+        definition.signal({ id: 'cond', type: i === 8 ? 'create' : 'ignore' });
+      }
+    });
+
+    Then('run completes', () => {
+      return completed;
+    });
+  });
+
   Scenario('With bound bad javascript condition', () => {
     let context, definition, callback;
     Given('a process matching scenario', async () => {
@@ -349,6 +535,69 @@ Feature('Conditional event', () => {
     Then('run is errored', async () => {
       const err = await errored;
       expect(err.content.error).to.have.property('source').with.property('content').with.property('id', 'start');
+    });
+  });
+
+  Scenario('external script resource', () => {
+    let definition;
+    Given('a source with external resource condition', async () => {
+      const source = `
+      <definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xmlns:js="http://paed01.github.io/bpmn-engine/schema/2017/08/bpmn">
+        <process id="theProcess" isExecutable="true">
+          <startEvent id="start" />
+          <manualTask id="task" />
+          <boundaryEvent id="cond" attachedToRef="task" cancelActivity="true">
+            <conditionalEventDefinition>
+              <condition xsi:type="tFormalExpression" language="js" js:resource="./external.js" />
+            </conditionalEventDefinition>
+          </boundaryEvent>
+          <endEvent id="end" />
+          <sequenceFlow id="flow0" sourceRef="start" targetRef="task" />
+          <sequenceFlow id="flow1" sourceRef="task" targetRef="end" />
+          <sequenceFlow id="flow2" sourceRef="cond" targetRef="end" />
+        </process>
+      </definitions>`;
+
+      const context = await testHelpers.context(source, {
+        extensions,
+        scripts: {
+          register(elm) {
+            if (elm.behaviour.resource === './external.js') {
+              return {
+                execute(_executionContext, callback) {
+                  return callback(null, { input: 3 });
+                },
+              };
+            }
+          },
+          getScript(_, elm) {
+            if (elm.behaviour.resource === './external.js') {
+              return {
+                execute(_executionContext, callback) {
+                  return callback(null, { input: 3 });
+                },
+              };
+            }
+          },
+        },
+      });
+
+      definition = new Definition(context);
+    });
+
+    let completed;
+    When('ran and condition is met', () => {
+      definition.environment.addService('conditionMet', () => {
+        return true;
+      });
+
+      completed = definition.waitFor('leave');
+      definition.run();
+    });
+
+    Then('run completes', () => {
+      return completed;
     });
   });
 });